+++ code/simulation_logic.cpp
@@ -22,7 +22,7 @@ bool simulation::runSimulationForOneParamSetting() {
     //
     std::cout << "Run simulation for a complete episode ..." << std::endl;
 
-    int n_tsteps = Global::get_n_timesteps();
+    unsigned long n_tsteps = Global::get_n_timesteps();
     struct tm* tm_start = Global::get_ts_start_tm();
     struct tm* tm_end   = Global::get_ts_end_tm();
     //time_t t_start = mktime(tm_start);
@@ -30,7 +30,7 @@ bool simulation::runSimulationForOneParamSetting() {
     struct tm* current_tm;
     bool sim_started = false; // gets true, if simulation range (as given by tm_start) has been reached
     // main loop
-    for (int ts = 1; ts <= n_tsteps; ts++) {
+    for (unsigned long ts = 1; ts <= n_tsteps; ts++) {
         // get current time as struct tm
         current_tm = global::time_localtime_str->at(ts - 1);
         // jump time steps if they are not inside the simulation range
@@ -63,7 +63,7 @@ bool simulation::runSimulationForOneParamSetting() {
 
 }
 
-bool simulation::oneStep(int ts) {
+bool simulation::oneStep(unsigned long ts) {
     //
     // Run one time step of the simulation
     // Return false if an error occurs during execution.
@@ -75,8 +75,8 @@ bool simulation::oneStep(int ts) {
     // loop over all control units:
     //    set new values and execute next actions
     ControlUnit*const* cuList = ControlUnit::GetArrayOfInstances();
-    const int nCUs = ControlUnit::GetNumberOfInstances();
-    for (int i = 0; i < nCUs; i++) {
+    const size_t nCUs = ControlUnit::GetNumberOfInstances();
+    for (size_t i = 0; i < nCUs; i++) {
         if (!cuList[i]->compute_next_value(ts))
             return false;
     }
@@ -91,11 +91,11 @@ bool simulation::oneStep(int ts) {
     // loop over all substations: compute new load values
     // and calculate total grid load
     Substation*const* subList = Substation::GetArrayOfInstances();
-    const int nSubst = Substation::GetNumberOfInstances();
+    const size_t nSubst = Substation::GetNumberOfInstances();
 
     if (output::substation_output != NULL) {
         *(output::substation_output) << ts << ","; // add timestep to output
-        for (int i = 0; i < nSubst; i++) {
+        for (size_t i = 0; i < nSubst; i++) {
             float current_station_load = subList[i]->calc_load();
             total_load += current_station_load;
             // stuff for output
@@ -141,7 +141,7 @@ bool simulation::runSimulationForAllVariations(int scenario_id) {
             for (auto& var_name_and_val : list_of_variables) {
 
                 if        (var_name_and_val.first.compare("expansion PV kWp static") == 0) {
-                    for (int i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
+                    for (size_t i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
                         cuList[i]->set_exp_pv_params_A(var_name_and_val.second);
                     cParamVals.exp_pv_kWp_static     = var_name_and_val.second;
                     cParamVals.exp_pv_kWp_static_set = true;
@@ -159,13 +159,13 @@ bool simulation::runSimulationForAllVariations(int scenario_id) {
                     cParamVals.exp_pv_kWp_per_m2_set = true;
 
                 } else if (var_name_and_val.first.compare("expansion BS P in kW")  == 0) {
-                    for (int i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
+                    for (size_t i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
                         cuList[i]->set_exp_bs_maxP_kW(var_name_and_val.second);
                     cParamVals.exp_bs_maxP_kW = var_name_and_val.second;
                     cParamVals.exp_bs_maxP_kW_set = true;
 
                 } else if (var_name_and_val.first.compare("expansion BS E in kWh") == 0) {
-                    for (int i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
+                    for (size_t i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
                         cuList[i]->set_exp_bs_maxE_kWh(var_name_and_val.second);
                     cParamVals.exp_bs_maxE_kWh = var_name_and_val.second;
                     cParamVals.exp_bs_maxE_kWh_set = true;
@@ -188,7 +188,7 @@ bool simulation::runSimulationForAllVariations(int scenario_id) {
                 if (cParamVals.exp_pv_kWp_per_m2_set)       kWp_per_m2 = cParamVals.exp_pv_kWp_per_m2;
                 if (cParamVals.exp_pv_min_kWp_roof_sec_set) min_kWp    = cParamVals.exp_pv_min_kWp_roof_sec;
                 if (cParamVals.exp_pv_max_kWp_roof_sec_set) max_kWp    = cParamVals.exp_pv_max_kWp_roof_sec;
-                for (int i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
+                for (size_t i = 0; i < ControlUnit::GetNumberOfInstances(); i++)
                     cuList[i]->set_exp_pv_params_B(kWp_per_m2, min_kWp, max_kWp);
             }
 
@@ -237,7 +237,7 @@ bool simulation::runSimulationForAllVariations(int scenario_id) {
     return true;
 }
 
-bool simulation::runSimulationFAVsAndSAC(float expansion_matrix_rel_freq[16][16], int expansion_matrix_abs_freq[16][16], int scenario_id) {
+bool simulation::runSimulationFAVsAndSAC(float expansion_matrix_rel_freq[16][16], long expansion_matrix_abs_freq[16][16], int scenario_id) {
     //
     // This function executes the expansion / adds sim. added components to selected CUs
     // and then calls runSimulationForAllVariations(int)
@@ -255,13 +255,13 @@ bool simulation::runSimulationFAVsAndSAC(float expansion_matrix_rel_freq[16][16]
         expansion::add_expansion_to_units(expansion_matrix_rel_freq, expansion_matrix_abs_freq, scenario_id, true);
         // 1.2a) add PV installations to all CUs, if they do not already have one
         ControlUnit*const* cuList = ControlUnit::GetArrayOfInstances();
-        const int nCUs = ControlUnit::GetNumberOfInstances();
-        for (int i = 0; i < nCUs; i++) {
+        const size_t nCUs = ControlUnit::GetNumberOfInstances();
+        for (size_t i = 0; i < nCUs; i++) {
             if (!cuList[i]->has_pv())
                 cuList[i]->add_exp_pv();
         }
         // 1.2b) add Battery to all CUs (that have a sim. added PV-installation), if they do not already have a battery
-        for (int i = 0; i < nCUs; i++) {
+        for (size_t i = 0; i < nCUs; i++) {
             if ( !cuList[i]->has_bs() && ( cuList[i]->get_exp_combi_bit_repr_sim_added() & expansion::MaskBS ) )
                 cuList[i]->add_exp_bs();
         }
@@ -273,14 +273,14 @@ bool simulation::runSimulationFAVsAndSAC(float expansion_matrix_rel_freq[16][16]
         }
         // 1.4) sort control units according to the SSR which has been computed in 1.2)
         // 1.4.a) Collect the SSR values
-        vector<pair<float, ControlUnit*>> ssr_cu_pair_vector( ControlUnit::GetNumberOfInstances() );
-        for (int i = 0; i < nCUs; i++) {
+        vector<pair<double, ControlUnit*>> ssr_cu_pair_vector( ControlUnit::GetNumberOfInstances() );
+        for (size_t i = 0; i < nCUs; i++) {
             ssr_cu_pair_vector.emplace_back(cuList[i]->get_SSR(), cuList[i]); // SSR, pointer to the object
         }
         // 1.4.b) Sort the SSR / CU-Pointer vactor
         sort(ssr_cu_pair_vector.begin(),
              ssr_cu_pair_vector.end(),
-             [](pair<float, ControlUnit*> a, pair<float, ControlUnit*> b) { return a.first > b.first; });
+             [](pair<double, ControlUnit*> a, pair<double, ControlUnit*> b) { return a.first > b.first; });
         //
         // 1.4.c) output metrics
         output::outputMetrics(true);
